// components/DynamicIslandBanner/interfaces/notification-layout.interface.ts
// Vendored from rit3zh/expo-dynamic-notifications @ 5de059a (MIT License). Imports rewritten from @/ alias to relative paths.
interface INotificationLayout {
  width: number;
  centerX: number;
  islandWidth: number;
  islandHeight: number;
  islandTop: number;
  islandBottom: number;
  islandCenterY: number;
  islandRadius: number;
  cardWidth: number;
  cardHeight: number;
  cardRadius: number;
  cardTop: number;
  cardLeft: number;
  cardCenterY: number;
  dropSize: number;
  neckWidth: number;
  canvasHeight: number;
}

export type { INotificationLayout };